requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.0.12
django-cleanup==5.2.0
django-crispy-forms==1.11.2
django-formtools==2.1
django-heroku==0.3.1
django-import-export==1.0.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==2.0.13
django-simple-captcha==0.5.6
django-social-share==2.2.1
django-xadmin==0.5.0
et-xmlfile==1.0.1
future==0.16.0
gunicorn==20.1.0
httplib2==0.9.2
idna==2.10
jdcal==1.3
lxml==4.6.3
odfpy==1.3.6
openpyxl==2.5.0
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.9.3
pycryptodome==3.12.0
python-dateutil==2.8.1
python-pptx==0.6.19
pytz==2021.1
PyYAML==5.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tablib==0.12.1
typing-extensions==3.10.0.0
unicodecsv==0.14.1
urllib3==1.26.5
whitenoise==5.2.0
wincertstore==0.2
xlrd==1.1.0
XlsxWriter==1.4.3
xlwt==1.3.0
